lab 4

+++ src/main/java/mk/ukim/finki/wp/wp_lab/web/controllers/CourseController.java
@@ -1,24 +1,21 @@
 package mk.ukim.finki.wp.wp_lab.web.controllers;
 
 import mk.ukim.finki.wp.wp_lab.model.Course;
-import mk.ukim.finki.wp.wp_lab.model.Student;
 import mk.ukim.finki.wp.wp_lab.model.Teacher;
 import mk.ukim.finki.wp.wp_lab.model.Type;
 import mk.ukim.finki.wp.wp_lab.service.CourseService;
+import mk.ukim.finki.wp.wp_lab.service.GradeService;
 import mk.ukim.finki.wp.wp_lab.service.StudentService;
 import mk.ukim.finki.wp.wp_lab.service.TeacherService;
-import org.checkerframework.checker.nullness.Opt;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.thymeleaf.context.WebContext;
 
-import javax.swing.text.html.Option;
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.TreeSet;
-import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("/courses")
@@ -27,11 +24,13 @@ public class CourseController {
     private final CourseService courseService;
     private final TeacherService teacherService;
 
+    private final GradeService gradeService;
     private final StudentService studentService;
 
-    public CourseController(CourseService courseService, TeacherService teacherService, StudentService studentService) {
+    public CourseController(CourseService courseService, TeacherService teacherService, GradeService gradeService, StudentService studentService) {
         this.courseService = courseService;
         this.teacherService = teacherService;
+        this.gradeService = gradeService;
         this.studentService = studentService;
     }
 
@@ -46,6 +45,7 @@ public class CourseController {
         return "listCourses";
     }
 
+
     @PostMapping("/add")
     public String save(@RequestParam String courseName,
                        @RequestParam String courseDescription,
@@ -53,7 +53,7 @@ public class CourseController {
                        @RequestParam Long teacherId,
                        @RequestParam(required = false) Long courseId) {
 
-        if (!Objects.isNull(courseId)) {
+        if (Objects.nonNull(courseId)) {
             courseService.edit(courseName, courseDescription, teacherId, type, courseId);
         } else {
             courseService.save(courseName, courseDescription, teacherId, type);
@@ -62,8 +62,13 @@ public class CourseController {
     }
 
     @DeleteMapping("/delete/{id}")
+    @Transactional
     public String deleteCourse(@PathVariable Long id) {
-        courseService.deleteCourseById(id);
+        Optional<Course> course = courseService.findCourseById(id);
+        if(course.isPresent()) {
+            gradeService.deleteAllByCourse(course.get());
+            courseService.deleteCourseById(id);
+        }
         return "redirect:/courses";
     }
 
@@ -77,14 +82,15 @@ public class CourseController {
 
     @GetMapping("/edit-form/{id}")
     public String getEditCoursePage(@PathVariable Long id, Model model) {
-        if (courseService.findCourseById(id).isPresent()) {
-            Course course = courseService.findCourseById(id).get();
+        Optional<Course> course = courseService.findCourseById(id);
+        if (course.isPresent()) {
             List<Teacher> teachers = teacherService.findAll();
             model.addAttribute("teachers", teachers);
-            model.addAttribute("course", course);
+            model.addAttribute("course", course.get());
             model.addAttribute("types", Type.toListOfStrings());
             return "add-course";
         }
+
         return "redirect:/courses?error=IllegalArgument";
     }
 }
